+++ compiler/src/jdk.graal.compiler/src/jdk/graal/compiler/core/common/GraalOptions.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2009, 2025, Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2009, 2026, Oracle and/or its affiliates. All rights reserved.
  * DO NOT ALTER OR REMOVE COPYRIGHT NOTICES OR THIS FILE HEADER.
  *
  * This code is free software; you can redistribute it and/or modify it
@@ -87,6 +87,10 @@ public final class GraalOptions {
 
     @Option(help = "Performs partial escape analysis and scalar replacement optimization.", type = OptionType.Expert)
     public static final OptionKey<Boolean> PartialEscapeAnalysis = new OptionKey<>(true);
+
+    @Option(help = "Performs statement level code duplication at control flow merges to " +
+                   "specialize code to branch values where possible.", type = OptionType.Expert)
+    public static final OptionKey<Boolean> OptDuplication = new OptionKey<>(true);
 
     @Option(help = "", type = OptionType.Debug)
     public static final OptionKey<Integer> EscapeAnalysisIterations = new OptionKey<>(2);